Add and configure prometheus exporter
This switches to the application starting with puma rather than rackup.

I've run into issues using the stock Prometheus::Middleware::Collector.
This creates a subclass of Prometheus::Middleware::Collector that expects
metrics to be defined for it, and if not, interfaces with a no-op.

* The configuration logic is encapsulated in the Metrics module.
* Metric definitions are moved into a yaml file.
* Metric recording from within the application via the Metrics() method.

For example:
```
duration = Benchmark.realtime do
  # access an API
end
Metrics(:api_response_duration_seconds) do |metric|
  metric.observe(duration, labels: {source: "some-api-server"})
end
```